requirements/requirements.txt
aliyun-python-sdk-core>=2.13.25
alibabacloud_sts20150401
importlib-metadata; python_version < "3.8"
numpy>=1.16.0
oss2>=2.8.0
pyodps>=0.11.0
pyyaml>=5.3.1
six>=1.15.0
marshmallow
marshmallow-oneofschema>=3.0.1
eas_prediction>=0.20
alibabacloud_tea_util>=0.3.6, <1.0.0, !=0.3.9
alibabacloud_tea_openapi>=0.3.3, <1.0.0
alibabacloud_openapi_util>=0.1.6, <1.0.0
alibabacloud_endpoint_util>=0.0.3, <1.0.0
alibabacloud-credentials>=0.3.2
Deprecated
addict
backoff<2.0.0
semantic-version
requests
tqdm
prompt-toolkit
pyOpenSSL>=23.0.0
aiohttp
